api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bleach==6.3.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
fastapi==0.135.2
h11==0.16.0
idna==3.11
kafka-python==2.3.0
kaggle==2.0.0
kagglesdk==0.1.16
mysql-connector-python==9.6.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-slugify==8.0.4
redis==7.4.0
requests==2.33.1
six==1.17.0
starlette==1.0.0
text-unidecode==1.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
webencodings==0.5.1
